[feature] New script that allows to select the default layout option.
This allow to have specialized firmwares that have a default layout different from the all pads vial firmware,
while keeping all other layout options available in case the user wants to customize its keyboard later on.
This will prevent the user to flash the all pad vial firmware for him to do so.